# backend-foryou/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForYou API Documentation",
        "description": "API documentation for the ForYou project",
        "contact": {
            "email": "admin@foryou.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cotizaciones": {
            "get": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Get list of cotizaciones",
                "description": "Returns list of cotizaciones",
                "operationId": "e0fb56c6e647aa0a32b74fa223c8eab7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cotizacion"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Create a new cotizacion",
                "description": "Creates a new cotizacion",
                "operationId": "1935ef0baa1d3ede5e6780b1800b4ab9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CotizacionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cotizacion created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cotizacion"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/cotizaciones/{id}": {
            "get": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Get cotizacion by ID",
                "description": "Returns a single cotizacion",
                "operationId": "e8f5a5732a12bde5581b022443cd333c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cotizacion to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cotizacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cotizacion not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Update an existing cotizacion",
                "description": "Updates a cotizacion",
                "operationId": "e6afd7cc7fc883b0f6b6dc9946072195",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cotizacion to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CotizacionInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cotizacion updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cotizacion"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cotizacion not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Delete a cotizacion",
                "description": "Deletes a single cotizacion",
                "operationId": "05d367e9c13af87638f76782f568a878",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of cotizacion to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Cotizacion not found"
                    }
                }
            }
        },
        "/api/cotizaciones/suggest-materials": {
            "post": {
                "tags": [
                    "Cotizaciones"
                ],
                "summary": "Suggest materials for a solicitud using AI",
                "description": "Provides AI-driven material suggestions based on a solicitud ID.",
                "operationId": "fbea7fe2c220d34191de0912c3c111d2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_solicitud"
                                ],
                                "properties": {
                                    "id_solicitud": {
                                        "description": "ID of the solicitud for which to suggest materials",
                                        "type": "integer",
                                        "format": "int64"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Material suggestions generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material suggestions generated successfully"
                                        },
                                        "solicitud_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "suggestions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "material_name": {
                                                        "type": "string",
                                                        "example": "PTR 1 1/2"
                                                    },
                                                    "quantity": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 20
                                                    },
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "m"
                                                    },
                                                    "notes": {
                                                        "type": "string",
                                                        "example": "Para marco de puerta"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Error generating material suggestions"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cotizacion": {
                "title": "Cotizacion",
                "required": [
                    "id_solicitud",
                    "costo_total"
                ],
                "properties": {
                    "id_cotizacion": {
                        "description": "ID de la cotización",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_solicitud": {
                        "description": "ID de la solicitud asociada",
                        "type": "integer",
                        "format": "int64"
                    },
                    "fecha_cot": {
                        "description": "Fecha de la cotización",
                        "type": "string",
                        "format": "date-time"
                    },
                    "costo_total": {
                        "description": "Costo total de la cotización",
                        "type": "number",
                        "format": "float"
                    },
                    "notas": {
                        "description": "Notas adicionales de la cotización",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CotizacionInput": {
                "title": "Cotizacion Input",
                "required": [
                    "id_solicitud",
                    "costo_total"
                ],
                "properties": {
                    "id_solicitud": {
                        "description": "ID de la solicitud asociada",
                        "type": "integer",
                        "format": "int64"
                    },
                    "costo_total": {
                        "description": "Costo total de la cotización",
                        "type": "number",
                        "format": "float"
                    },
                    "notas": {
                        "description": "Notas adicionales de la cotización",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}